+++ src/main/java/com/example/thedayoftoday/app/CalendarController.java
@@ -1,8 +1,10 @@
 package com.example.thedayoftoday.app;
 
+import com.example.thedayoftoday.domain.dto.diary.DiaryEntryResponseDto;
+import com.example.thedayoftoday.domain.dto.calendar.MonthColorsResponseDto;
+import com.example.thedayoftoday.domain.dto.calendar.SentimentalAnalysisListResponseDto;
 import com.example.thedayoftoday.domain.security.CustomUserDetails;
 import com.example.thedayoftoday.domain.service.CalendarService;
-import com.example.thedayoftoday.domain.dto.*;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
