+++ src/main/java/com/example/thedayoftoday/app/DiaryController.java
@@ -1,7 +1,7 @@
 package com.example.thedayoftoday.app;
 
-import com.example.thedayoftoday.domain.dto.DiaryBasicResponseDto;
-import com.example.thedayoftoday.domain.dto.DiaryRequestDto;
+import com.example.thedayoftoday.domain.dto.diary.DiaryBasicResponseDto;
+import com.example.thedayoftoday.domain.dto.diary.DiaryRequestDto;
 import com.example.thedayoftoday.domain.dto.conversation.ConversationResponseDto;
 import com.example.thedayoftoday.domain.entity.DiaryMood;
 import com.example.thedayoftoday.domain.security.CustomUserDetails;
