+++ src/main/java/com/example/thedayoftoday/domain/dto/calendar/MonthColorsResponseDto.java
@@ -1,4 +1,4 @@
-package com.example.thedayoftoday.domain.dto;
+package com.example.thedayoftoday.domain.dto.calendar;
 
 import java.util.Map;
