+++ src/main/java/com/example/thedayoftoday/domain/dto/calendar/SentimentalAnalysisRequestDto.java
@@ -1,6 +1,6 @@
-package com.example.thedayoftoday.domain.dto;
+package com.example.thedayoftoday.domain.dto.calendar;
 
-public record SentimentalAnalysisResponseDto(
+public record SentimentalAnalysisRequestDto(
         String analysisMoodName,
         String analysisMoodColor,
         String analysisContent
