+++ src/main/java/com/example/thedayoftoday/domain/dto/calendar/SentimentalAnalysisResultDto.java
@@ -1,4 +1,4 @@
-package com.example.thedayoftoday.domain.dto;
+package com.example.thedayoftoday.domain.dto.calendar;
 
 import com.example.thedayoftoday.domain.entity.enumType.MoodMeter;
