+++ src/main/java/com/example/thedayoftoday/domain/dto/diary/DiaryBasicResponseDto.java
@@ -1,4 +1,4 @@
-package com.example.thedayoftoday.domain.dto;
+package com.example.thedayoftoday.domain.dto.diary;
 
 public record DiaryBasicResponseDto(String title,String content) {
 }
