+++ src/main/java/com/example/thedayoftoday/domain/dto/diary/DiaryEntryResponseDto.java
@@ -1,4 +1,4 @@
-package com.example.thedayoftoday.domain.dto;
+package com.example.thedayoftoday.domain.dto.diary;
 
 import java.util.List;
