+++ src/main/java/com/example/thedayoftoday/domain/dto/login/LoginRequestDto.java
@@ -1,4 +1,4 @@
-package com.example.thedayoftoday.domain.dto;
+package com.example.thedayoftoday.domain.dto.login;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotNull;
