+++ src/main/java/com/example/thedayoftoday/domain/dto/setting/UserSettingDto.java
@@ -1,4 +1,4 @@
-package com.example.thedayoftoday.domain.dto;
+package com.example.thedayoftoday.domain.dto.setting;
 
 public record UserSettingDto(String name, String email, String phoneNumber) {
 }
