+++ src/main/java/com/example/thedayoftoday/domain/dto/weeklyAnalysis/WeeklyTitleFeedbackResponseDto.java
@@ -1,4 +1,4 @@
-package com.example.thedayoftoday.domain.dto;
+package com.example.thedayoftoday.domain.dto.weeklyAnalysis;
 
 public record WeeklyTitleFeedbackResponseDto(String title, String feedback) {
 }
