+++ src/main/java/com/example/thedayoftoday/domain/service/CalendarService.java
@@ -1,6 +1,10 @@
 package com.example.thedayoftoday.domain.service;
 
-import com.example.thedayoftoday.domain.dto.*;
+import com.example.thedayoftoday.domain.dto.diary.DiaryContentDto;
+import com.example.thedayoftoday.domain.dto.diary.DiaryEntryResponseDto;
+import com.example.thedayoftoday.domain.dto.calendar.MonthColorsResponseDto;
+import com.example.thedayoftoday.domain.dto.calendar.SentimentalAnalysisListResponseDto;
+import com.example.thedayoftoday.domain.dto.calendar.SentimentalAnalysisResultDto;
 import com.example.thedayoftoday.domain.entity.Diary;
 import com.example.thedayoftoday.domain.entity.enumType.MoodMeter;
 import com.example.thedayoftoday.domain.repository.DiaryRepository;
