+++ src/test/java/com/example/thedayoftoday/domain/CalendarServiceTest.java
@@ -1,6 +1,9 @@
 package com.example.thedayoftoday.domain;
 
-import com.example.thedayoftoday.domain.dto.*;
+import com.example.thedayoftoday.domain.dto.diary.DiaryEntryResponseDto;
+import com.example.thedayoftoday.domain.dto.calendar.MonthColorsResponseDto;
+import com.example.thedayoftoday.domain.dto.calendar.SentimentalAnalysisListResponseDto;
+import com.example.thedayoftoday.domain.dto.calendar.SentimentalAnalysisResultDto;
 import com.example.thedayoftoday.domain.entity.Diary;
 import com.example.thedayoftoday.domain.entity.DiaryMood;
 import com.example.thedayoftoday.domain.repository.DiaryRepository;
